revert(sessions): settle no longer stops background work
Cutting the second and last slice of settle teardown. Six review rounds, every
one of them finding a real defect in this specific mechanism:

- an unsettle path that skipped the resume (x3, each a route the previous fix
  had not traced),
- settling mid-turn tearing down nothing while the caller still wrote the marker,
- the RPC operator bridge never receiving the teardown control,
- a stop count that could not be kept honest,
- and finally an activity guard that reads lastActivityAt — which is backed by
  last_output_at, a column clearTurnStartMarkers never writes. The guard I added
  last round provably cannot fire.

The shape is now unambiguous. Teardown is async; settled_at is written and
cleared from seven places. A teardown-then-write settle races real activity, and
the failure is not one-sided: a user starting a turn during a provider stop call
gets their background work stopped AND no settle. Every guard against it either
read a column turn-start does not update, or had to be repeated identically at
each settle entry point (settleTerminalSession, bulk registry, both IPC
handlers, both sync commands, PR auto-settlement, the CTO tool).

Doing this correctly needs a synchronous lifecycle revision that teardown can be
serialized against — a different change, designed as one, not a wrapper around
the existing write. Shipping the half-working version is worse than the status
quo, which is the one thing the brief specifically warned about.

What ships instead is the half that has been stable since iteration 2 and is
what a user actually sees: live background work promoted into the canonical
phase across every glanceable surface, the working/monitoring denylist,
cross-runtime generalization, running-count subagent badges, and the archive
port-lease ordering fix — archive remains the lifecycle path that does stop
processes, and its ordering bug is fixed.

Removed: sessionMachineryTeardown, stopBackgroundWork, the activity guard, and
the teardown calls in every settle entry point. runtimeBackgroundWork stays; it
is the surfacing half.
