+++ apps/ade-cli/src/bootstrap.ts
@@ -1434,8 +1434,6 @@ export async function createAdeRuntime(args: {
   const prMergeAutoSettlementService = createPrMergeAutoSettlementService({
     db,
     sessionService,
-    agentChatService,
-    logger,
     emitEvent: emitPrEvent,
   });
