+++ apps/desktop/src/main/services/adeActions/registry.ts
@@ -126,7 +126,6 @@ import {
   parseWakeReason,
 } from "../sessions/sessionRequestValidation";
 import { settleTerminalSession } from "../sessions/settleTerminalSession";
-import { stopSettledSessionMachinery } from "../sessions/sessionMachineryTeardown";
 import {
   getSessionLifecycleSettings,
   setSessionLifecycleSettings,
@@ -2127,7 +2126,6 @@ function buildSessionDomainService(runtime: AdeRuntime): OpaqueService | null {
         sessionService,
         agentChatService: runtime.agentChatService,
         ptyService: runtime.ptyService,
-        logger: runtime.logger,
       })) {
         throw new Error(`Session '${sessionId}' was not found.`);
       }
@@ -2148,13 +2146,6 @@ function buildSessionDomainService(runtime: AdeRuntime): OpaqueService | null {
     },
     // Bulk settle/unsettle for renderer surfaces on remote-bound projects
     // (mirrors deleteSession's generic trust posture).
-    //
-    // Deliberately NOT an `async` function: argument validation below must keep
-    // throwing SYNCHRONOUSLY, the way it did before settle grew a teardown
-    // step. Marking the whole action `async` silently converts every one of
-    // those guards into a rejected promise, which changes the contract for any
-    // caller that does not await — so only the success path is async, returned
-    // as an explicit promise from a sync body.
     settleSessions: (args?: unknown) => {
       const record = readObjectActionArg(args, "session.settleSessions");
       const sessionIds = Array.isArray(record.sessionIds)
@@ -2177,17 +2168,7 @@ function buildSessionDomainService(runtime: AdeRuntime): OpaqueService | null {
           "session.settleSessions does not dismiss pending input; use session.settleSession for a single session.",
         );
       }
-      return (async () => {
-        await stopSettledSessionMachinery(
-          {
-            sessionService,
-            agentChatService: runtime.agentChatService,
-            logger: runtime.logger,
-          },
-          sessionIds,
-        );
-        return sessionService.settleSessions(sessionIds);
-      })();
+      return sessionService.settleSessions(sessionIds);
     },
     unsettleSessions: (args?: unknown) => {
       const record = readObjectActionArg(args, "session.unsettleSessions");
