+++ apps/desktop/src/main/services/ipc/registerIpc.ts
@@ -51,7 +51,6 @@ import {
   parseWakeReason,
 } from "../sessions/sessionRequestValidation";
 import { settleTerminalSession } from "../sessions/settleTerminalSession";
-import { stopSettledSessionMachinery } from "../sessions/sessionMachineryTeardown";
 import {
   getSessionLifecycleSettings,
   setSessionLifecycleSettings,
@@ -6967,7 +6966,6 @@ export function registerIpc({
         sessionService: ctx.sessionService,
         agentChatService: ctx.agentChatService,
         ptyService: ctx.ptyService,
-        logger: ctx.logger,
       });
       if (!settled) throw new Error(`Session '${sessionId}' was not found.`);
     },
@@ -6988,14 +6986,9 @@ export function registerIpc({
     async (_event, arg: { sessionIds?: unknown }): Promise<string[]> => {
       const ctx = ensureSessionContext();
       if (!Array.isArray(arg?.sessionIds)) throw new Error("Session ids are required.");
-      const sessionIds = arg.sessionIds.filter(
-        (sessionId): sessionId is string => typeof sessionId === "string",
+      return ctx.sessionService.settleSessions(
+        arg.sessionIds.filter((sessionId): sessionId is string => typeof sessionId === "string"),
       );
-      await stopSettledSessionMachinery(
-        { sessionService: ctx.sessionService, agentChatService: ctx.agentChatService, logger: ctx.logger },
-        sessionIds,
-      );
-      return ctx.sessionService.settleSessions(sessionIds);
     },
   );
