+++ apps/desktop/src/main/services/prs/prAsync.test.ts
@@ -912,47 +912,6 @@ describe("prMergeAutoSettlementService", () => {
     expect(settleSessionsWithOutcome).toHaveBeenCalledTimes(2);
   });
 
-  it("stops the merged session's machinery before filing it", async () => {
-    // This path deliberately bypasses the settlement blockers, so it is the
-    // one most likely to file a session that is still running something. Before
-    // teardown reached it, the merged lane's monitors kept polling and woke the
-    // thread hours after the PR had landed.
-    const db = createMemoryDb();
-    const order: string[] = [];
-    const settleSessionsWithOutcome = vi.fn((ids: string[]) => {
-      order.push("settle");
-      return ids;
-    });
-    const stopBackgroundWork = vi.fn(async () => {
-      order.push("stop");
-      return { skippedActiveTurn: false };
-    });
-    const rows = [{ id: "chat-live", toolType: "claude-chat", archivedAt: null, settledAt: null }];
-    const service = createPrMergeAutoSettlementService({
-      db: db as any,
-      sessionService: {
-        list: vi.fn(() => rows),
-        get: vi.fn((id: string) => rows.find((row) => row.id === id) ?? null),
-        settleSessionsWithOutcome,
-      } as any,
-      agentChatService: {
-        stopBackgroundWork,
-      } as any,
-      emitEvent: vi.fn(),
-    });
-
-    await service.processSnapshot({
-      prs: [createSummary({ state: "open" })],
-      polledAt: "2026-03-24T12:00:00.000Z",
-    });
-    await service.processSnapshot({
-      prs: [createSummary({ state: "merged", mergedAt: "2026-03-24T12:01:00.000Z" })],
-      polledAt: "2026-03-24T12:01:30.000Z",
-    });
-
-    expect(stopBackgroundWork).toHaveBeenCalledWith({ sessionId: "chat-live" });
-    expect(order).toEqual(["stop", "settle"]);
-  });
 
   it("does not re-settle after reactivation, but settles for a later PR", async () => {
     const db = createMemoryDb();
