+++ apps/desktop/src/main/services/sessions/settleTerminalSession.ts
@@ -3,7 +3,6 @@ import type { createPtyService } from "../pty/ptyService";
 import type { createSessionService } from "./sessionService";
 import type { SessionSettleSource } from "../../../shared/types";
 import { isChatToolType } from "./chatSessionProjection";
-import { stopSettledSessionMachinery } from "./sessionMachineryTeardown";
 
 export type SettleTerminalSessionOptions = {
   outcome?: string;
@@ -46,21 +45,12 @@ export async function dismissPendingInputBeforeSettle(args: {
   return true;
 }
 
-/**
- * Settle a session AND stop the machinery it owns.
- *
- * The teardown runs before the column write so a settle can never report
- * success while its monitors are still armed. It is best-effort — see
- * `stopSettledSessionMachinery` — so a provider that cannot be reached delays
- * nothing and blocks nothing.
- */
 export async function settleTerminalSession(args: {
   sessionId: string;
   opts?: SettleTerminalSessionOptions;
   sessionService: ReturnType<typeof createSessionService>;
   agentChatService?: ReturnType<typeof createAgentChatService> | null;
   ptyService?: ReturnType<typeof createPtyService> | null;
-  logger?: { warn: (message: string, meta?: Record<string, unknown>) => void } | null;
 }): Promise<boolean> {
   if (args.opts?.dismissPendingInput === true) {
     const dismissed = await dismissPendingInputBeforeSettle({
@@ -70,34 +60,6 @@ export async function settleTerminalSession(args: {
       ptyService: args.ptyService ?? null,
     });
     if (!dismissed) return false;
-  }
-
-  // Teardown awaits provider stop calls that can take seconds, and a user can
-  // start a new turn inside that window. `clearTurnStartMarkers` would clear a
-  // settle marker that does not exist yet, and this write would then file the
-  // freshly-active session as settled. Snapshot the activity stamp first and
-  // refuse to settle over work that arrived while we were stopping things —
-  // real activity outranks a settle request that predates it.
-  const activityBeforeTeardown = args.sessionService.get(args.sessionId)?.lastActivityAt ?? null;
-
-  await stopSettledSessionMachinery(
-    {
-      sessionService: args.sessionService,
-      agentChatService: args.agentChatService ?? null,
-      logger: args.logger ?? null,
-    },
-    [args.sessionId],
-  );
-
-  const after = args.sessionService.get(args.sessionId);
-  if (!after) return false;
-  if ((after.lastActivityAt ?? null) !== activityBeforeTeardown) {
-    args.logger?.warn("session_teardown.settle_skipped_new_activity", {
-      sessionId: args.sessionId,
-    });
-    // The row exists and the request was honoured; it simply woke, so it is not
-    // settled. Reporting false here would surface a spurious "not found".
-    return true;
   }
 
   return args.sessionService.settleSession(
